requirements.txt
python-dotenv
langchain
langchain-community
psycopg2-binary
SQLAlchemy 
pypdf
langchain-openai
docx2txt
numpy